Fix arg type of operator[]

+++ Matrix.cpp
@@ -50,7 +50,7 @@ struct Matrix
     }
 
     // Return to the x row. We count from 0, which is not common in math.
-    vector<double>& operator[](unsigned int x)
+    vector<double>& operator[](size_t x)
     {
         return data[x];
     }
